+++ mayan/__init__.py
@@ -1,7 +1,7 @@
 __title__ = 'Mayan EDMS'
 __version__ = '4.7dev0'
 __build__ = 0x040602
-__build_string__ = 'v4.5.6-479-gaf6d9044725_Sat Mar 9 18:45:38 2024 -0400'
+__build_string__ = 'v4.5.6-481-g79283b4f715_Sun Mar 10 01:47:11 2024 -0400'
 __django_version__ = '4.2'
 __author__ = 'Roberto Rosario'
 __author_email__ = 'roberto.rosario@mayan-edms.com'