+++ mayan/apps/converter/dependencies.py
@@ -27,7 +27,7 @@ PythonDependency(
     version_string='==10.2.0'
 )
 PythonDependency(
-    module=__name__, name='pypdf', version_string='==3.17.4'
+    module=__name__, name='pypdf', version_string='==4.1.0'
 )
 PythonDependency(
     module=__name__, name='qrcode', version_string='==7.4.2'