+++ mayan/apps/storage/dependencies.py
@@ -7,10 +7,10 @@ PythonDependency(
     module=__name__, name='django-storages', version_string='==1.14.2'
 )
 PythonDependency(
-    module=__name__, name='extract-msg', version_string='==0.48.0'
+    module=__name__, name='extract-msg', version_string='==0.48.2'
 )
 PythonDependency(
-    module=__name__, name='google-cloud-storage', version_string='==2.10.0'
+    module=__name__, name='google-cloud-storage', version_string='==2.15.0'
 )
 PythonDependency(
     module=__name__, name='pycryptodome', version_string='==3.20.0'